+++ src/components/POLine/LineLinkedInstances/LineLinkedInstances.test.js
@@ -1,22 +1,31 @@
 import React from 'react';
 import { act, render, screen } from '@testing-library/react';
+import user from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
+import { useLocation, useHistory } from 'react-router';
 
 import { instance } from '../../../../test/jest/fixtures';
+import { location, history } from '../../../../test/jest/routerMocks';
+
 import InstancePlugin from '../Item/InstancePlugin';
 import { ACCORDION_ID } from '../const';
 import { LineLinkedInstances } from './LineLinkedInstances';
 import { useLinkedInstances, useTitleMutation } from './hooks';
 
-jest.mock('./hooks', () => ({
-  useLinkedInstances: jest.fn(),
-  useTitleMutation: jest.fn().mockReturnValue({}),
+jest.mock('react-router', () => ({
+  ...jest.requireActual('react-router'),
+  useHistory: jest.fn(),
+  useLocation: jest.fn(),
 }));
 jest.mock('@folio/stripes/components', () => ({
   ...jest.requireActual('@folio/stripes/components'),
   Loading: jest.fn().mockReturnValue('Loading'),
 }), { virtual: true });
 jest.mock('../Item/InstancePlugin', () => jest.fn().mockReturnValue('InstancePlugin'));
+jest.mock('./hooks', () => ({
+  useLinkedInstances: jest.fn(),
+  useTitleMutation: jest.fn().mockReturnValue({}),
+}));
 
 const renderLineLinkedInstances = ({ line = {}, toggleSection = jest.fn(), labelId = 'labelId' }) => (render(
   <LineLinkedInstances
@@ -31,6 +40,9 @@ describe('LineLinkedInstances', () => {
   beforeEach(() => {
     useLinkedInstances.mockClear();
     useTitleMutation.mockClear();
+
+    useHistory.mockClear().mockReturnValue(history);
+    useLocation.mockClear().mockReturnValue(location);
   });
 
   it('displays table with records with fetched instances', () => {
@@ -92,6 +104,32 @@ describe('LineLinkedInstances', () => {
       });
 
       expect(mutateTitle).toHaveBeenCalled();
+    });
+  });
+
+  describe('Row select', () => {
+    it('should not open receiving title edit form when instance title is clicked', () => {
+      const linkedInstance = { title: 'ABA', id: 'instanceUid', contributors: 'Mark' };
+
+      useLinkedInstances.mockReturnValue({ isLoading: false, linkedInstances: [linkedInstance] });
+
+      renderLineLinkedInstances({ line: { instanceId: 'instanceUid' } });
+
+      user.click(screen.getByText(linkedInstance.title));
+
+      expect(history.push).not.toHaveBeenCalled();
+    });
+
+    it('should open receiving title edit form when instance row is clicked', () => {
+      const linkedInstance = { title: 'ABA', id: 'instanceUid', contributors: 'Mark', receivingTitle: { id: 'titleId' } };
+
+      useLinkedInstances.mockReturnValue({ isLoading: false, linkedInstances: [linkedInstance] });
+
+      renderLineLinkedInstances({ line: { instanceId: 'instanceUid' } });
+
+      user.click(screen.getByText(linkedInstance.contributors));
+
+      expect(history.push).toHaveBeenCalled();
     });
   });
 });