+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add constraint for POL: physical POL should contain only Physical element and electronic only E-resource. Refs UIOR-825.
 * Add "Package titles" accordion to package POL view. Refs UIOR-831.
 * Allow user to add Package titles from POL view. Refs UIOR-832.
+* Allow user to select instance for Title AND edit title from POL view. Refs UIOR-833.
 
 ## [3.0.2](https://github.com/folio-org/ui-orders/tree/v3.0.2) (2021-11-05)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.1...v3.0.2)